Add SPoILeR (ECCV 2026) to the MultimodalStudio repository Release the code of "Learning Spectral and Polarimetric Clues for One-to-Multimodal Novel View Synthesis" on top of the MultimodalStudio framework, which the two works now share. Framework: - new modules: dictionary fields, latent radiance model, latent estimators,   per-scene module replication, multi-scene and fine-tuning pipelines - the InstantNGP geometry becomes the standard of the dictionary_fields   family, replacing DictionarySDFField; the registry exposes 21 methods - shared modules updated accordingly (encodings, losses, evaluator,   camera optimizers, data manager, ...) Configurations: - confs/ split into confs/multimodalstudio/ and confs/spoiler/ - 24 SPoILeR configurations: multi-scene pre-training, per-scene   fine-tuning, few-view settings and loss ablations - confs/template.yaml extended with the options added by SPoILeR Tools and documentation: - scripts/evaluate_average_metrics.py handles both single-scene and   multi-scene runs, and fixes the demosaicing call - README: both papers, SPoILeR training workflow and results - docs: pipelines section and corrected configuration example Licensing: - LICENSE_FACTOR_FIELDS.txt reports the MIT license of the Factor Fields   project and the list of files it applies to

+++ src/model_components/renderers.py
@@ -225,7 +225,7 @@ class SemanticRenderer(BaseRenderer):
         weights: TensorType["bs":..., "num_samples", 1],
     ) -> TensorType["bs":..., "num_classes"]:
         """Calculate semantics along the ray."""
-        sem = torch.sum(weights * semantics, dim=-2)
+        sem = torch.sum(weights * semantics, dim=-2).to(semantics.dtype)
         return sem
 
 class NormalsRenderer(BaseRenderer):
